Converted TabHeader to Array members.
